+++ src/plugins/router/index.js
@@ -1,0 +1,8 @@
+export { Router, useRouter } from './Router.jsx'
+export { default as Route } from './Route.jsx'
+export { default as Link } from './Link.jsx'
+export { default as OffCanvas } from './OffCanvas.jsx'
+export { useFocusOnMount } from './useFocusOnMount.js'
+export { useReturnFocus } from './useReturnFocus.js'
+export { useMediaQuery } from './useMediaQuery.js'
+export { useFocusTrap } from './useFocusTrap.js'